refactor: extract WorkerRegistry owning the worker-LLM cache (R6 worker sub-seam)
